feat: 아이콘 버튼 분리 (#27)
