+++ src/components/common/Button/Button.style.ts
@@ -1,0 +1,29 @@
+import { cva } from 'class-variance-authority';
+
+export const button = cva(
+  'rounded-[12px] w-full h-[5.6rem] flex justify-center items-center',
+  {
+    variants: {
+      textSize: {
+        sm: 'title1',
+        lg: 'head1',
+      },
+      backgroundColor: {
+        dark: 'bg-button-color',
+        light: 'bg-background-02',
+        gray: 'bg-gray-600',
+        disabled: 'bg-gray-300',
+      },
+      textColor: {
+        black: 'text-text',
+        gray: 'text-gray-400',
+        white: 'text-white',
+      },
+    },
+    defaultVariants: {
+      textSize: 'lg',
+      backgroundColor: 'dark',
+      textColor: 'white',
+    },
+  },
+);
